+++ CJBCheatsMenu/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Updated for Stardew Valley 1.5, including...
   * split-screen mode and UI scaling;
+  * new fast machines (bone mill, coffee maker, deconstructor, geode crusher, heavy tapper, ostrich incubator, and solar panel);
   * key to the town wallet item;
   * island warps.